refactor: move sign handler to qsign.go
